register all required services, move services to correct folders
