+++ Branta/V2/Extensions/ConfigurationExtension.cs
@@ -17,8 +17,9 @@ public static class ConfigurationExtension
         {
             services.AddSingleton(Options.Create(defaultOptions));
         }
+        services.AddScoped<IAesEncryption, AesEncryptionService>();
         services.AddSingleton<ISecretGenerator, GuidSecretGenerator>();
-        services.AddScoped<BrantaClient>();
+        services.AddScoped<IBrantaClient, BrantaClient>();
         services.AddScoped<IBrantaService, BrantaService>();
 
         return services;
